+++ Packages/BrandonUtils/Standalone/Strings/PrettificationFlags.cs
@@ -1,0 +1,9 @@
+﻿using System;
+
+namespace BrandonUtils.Standalone.Strings {
+    [Flags]
+    public enum PrettificationFlags {
+        None = 0,
+        IncludeTypeLabels
+    }
+}